Always show all rows in table view of browse tab
Always show all rows of the current table in the browse tab of the main
window instead of limiting it to a 1000 per page.
Change the functionality of the buttons and inputs to change the page
accordingly.

Limiting the number of rows to a 1000 was probably done for performance
reasons but is a killer in terms of usability. And as we'll probably
move sooner or later to a model-view architecture that loads data on
demand, the performance when opening large tables will be more than
restored anyway.