Update the about dialog
Fix the link to the Qt project page. Also link to their page for license
information instead of pointing at a non-existant file in our
repository. This fixed issue #5.

Mention the Silk icon set in the about dialog and link to its page. This
should also make the hard-to-find CREDITS file dispensable.

Generally redesign the about dialog to use layouts for its design.