Fix splitter between code editor and result view in Execute SQL tab
There was a supposed to be a splitter between the SQL code editor widget
and the result table view widget in the SQL execution areas in the
Execute SQL tab. A splitter like this allows resizing the widgets around
it. Apparently this was broken some time ago (probably by me) and is
fixed by this commit.

See #380 which is partially fixed by this.