Fix opening of edit dialog when pressing the return key
Pressing the return key opens the edit dialog or moves focus to the edit
dock (whichever is used). However, when using the edit *dialog* doing
inline editing and then hitting the return key opens the edit dialog
instead of just saving the changes. This is changed by this commit.

See issue #480.