Add SQLCipher version string to the About Dialog
This isn't fully worked out yet, as it appears to need an open
database connection for the string retrieval to work.  Without
an open database the field is left blank.

Not sure if it's something caused by us, or a requirement from
the underlying SQLCipher library itself.

Either way, we'll need to find a solution to that for this to
be useful.