Stop stealing focus() from the Browse Data tab to the Edit Cell
This is to address #662.  Prior to this patch, when in "Window"
mode the Edit Cell steals focus from the Browse Data tab.  This
means keyboard navigation in the Browse Data tab is impossible. :(

This patch fixes that issue, allowing keyboard navigation to work
correctly in the Browse Data tab again.