Load the full cell data when trying to edit the cell content
In 11bff0ded839c7d4b6b5eea6118ac44938515ce5 a settings option was
introduced in order to limit the number of characters loaded into each
cell in the Browse Data tab.

However, when editing the cell you wouldn't see the full data, i.e.
rendering the data incomplete and/or cropping it when it's being edited
anyway. This commit fixes this behaviour: when editing a cell (either
in-line or using the edit dialog/dock) the full data for this particular
cell is loaded and shown.