Fix executing current SQL running incorrect line
Previously, it was running the query at the current cursor position, but that
proved to be confusing. Now, it runs the SQL that contains the start of the
current line. So having:

SELECT * FROM table1;
SELECT * FROM table2;

and the cursor either before the first semicolon, either after it, the first
SQL will be run, and not the second one as before.