Fix altering tables with multiple indices
This hopefully prevents an error message from appearing when trying to
edit a table with multiple indices referring to it. While restoring
these indices the SQL commands wouldn't be separated by semicolons and
that might have caused problems.

Also trim the SQL commands to remove any excessive whitespace that might
be in them.

See issue #771.