Fix renaming of table columns when there are foreign keys constraints
When editing a table (e.g. renaming a column) and there is another table
referencing the edited table in a foreign key, this could cause problems
when foreign key enforcement is enabled. We tried to circumnavigate
these issues by temporarily disabling foreign keys but that wouldn't
work. This commit tries to implement a better way of achieving the same
goal.