Store contraints in a hash structure instead of a map
Our key type here is a vector of fields and vectors aren't less than
comparable by default (and I have no idea how to implement a sane
operator<() for them). But because maps require less than comparability,
change the map to a hash which only requires an operator==() which a
vector should have by default.

No idea how this ever worked with a map?!

See issue #854.