Fix adding primary keys to tables in the Edit Table dialog
Adding a primary key to a table that didn't had a primary key before
would have added the column to the sqlb::Table::primaryKey::emptyFieldVector
variable (which is static!) instead of actually creating a new primary
key.

This way the primary key wouldn't be created as a new table constraint.

See issue #918.